fix: 回调失效与内存安全修复（Commit5 问题单 7 项）
修复 Commit5 问题单 7 项：范围检查恒真导致回调全部失效、持锁调用回调、
SSAP 被动配对跨线程 UAF 读、零长度读 double free、结构体长度 uint8_t 截断、
多读错误码未同步、流音量数据越界读。

1. ani_nearlink_manager_callback.cpp：OnStateChanged 范围检查边界修正（STATE_TURNING_ON~STATE_TURN_HALF），5 处回调改为锁内拷贝快照、锁外调用（问题1/2）
2. ssaps_server.c：SSAPS_ReadMultiHandleItem 失败分支补 result->errorCode 同步（问题3）
3. ssaps_server_find.c：GetStructureInfoSize descriptorTypeListLen 改 uint16_t 防截断少报 256 字节（问题4）
4. SleAdapter.cpp/h：ServiceSsapCallback 继承 enable_shared_from_this，回调捕获 shared_ptr 保活防跨线程 UAF（问题5）
5. hid_stm.c：零长度读先校验后释放，释放后置 NULL 防 double free（问题6）
6. mcp_volume_client.c：McpParseStreamVolumeData 解析前校验 dataLen>=1 防越界读（问题7）

+++ services/stack/src/cp/bsl/sle/servm/ssap/src/ssaps_server_find.c
@@ -1318,7 +1318,8 @@ static uint32_t GetStructureInfoSize(SSAP_FindStructureInfo_S *info)
     uint8_t operationLen = GetOperationLenByItemType(info->itemType);
 
     uint8_t descriptorCountLen = GetDescriptorCountLenByItemType(info->itemType);
-    uint8_t descriptorTypeListLen = 0;
+    // descriptorCountLen + descriptorCount 最大可达 256（1 + 255），用 uint16_t 避免截断少报长度
+    uint16_t descriptorTypeListLen = 0;
     if (descriptorCountLen != 0) {
         descriptorTypeListLen = descriptorCountLen + info->descriptorCount;
     }
